refactor: organize event v2 functions
