+++ pkg/skaffold/event/v2/event.go
@@ -24,11 +24,9 @@ import (
 
 	//nolint:golint,staticcheck
 	"github.com/golang/protobuf/jsonpb"
-	pbuf "github.com/golang/protobuf/proto"
 	"google.golang.org/protobuf/types/known/timestamppb"
 
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/constants"
-	sErrors "github.com/GoogleContainerTools/skaffold/pkg/skaffold/errors"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/schema/util"
 	"github.com/GoogleContainerTools/skaffold/proto/enums"
 	proto "github.com/GoogleContainerTools/skaffold/proto/v2"
@@ -93,11 +91,6 @@ func GetIteration() int {
 	return handler.iteration
 }
 
-func GetState() (*proto.State, error) {
-	state := handler.getState()
-	return state, nil
-}
-
 func ForEachEvent(callback func(*proto.Event) error) error {
 	return handler.forEachEvent(callback)
 }
@@ -106,51 +99,17 @@ func ForEachApplicationLog(callback func(*proto.Event) error) error {
 	return handler.forEachApplicationLog(callback)
 }
 
-func Handle(event *proto.Event) error {
-	if event != nil {
-		handler.handle(event)
+func (ev *eventHandler) forEachEvent(callback func(*proto.Event) error) error {
+	// Unblock call to `WaitForConnection()`
+	select {
+	case handler.wait <- true:
+	default:
 	}
-	return nil
+	return ev.forEach(&ev.eventListeners, &ev.eventLog, &ev.logLock, callback)
 }
 
-// WaitForConnection will block execution until the server receives a connection
-func WaitForConnection() {
-	<-handler.wait
-}
-
-func (ev *eventHandler) getState() *proto.State {
-	ev.stateLock.Lock()
-	// Deep copy
-	state := pbuf.Clone(ev.state).(*proto.State)
-	ev.stateLock.Unlock()
-
-	return state
-}
-
-func (ev *eventHandler) log(event *proto.Event, listeners *[]*listener, log *[]*proto.Event, lock sync.Locker) {
-	lock.Lock()
-
-	for _, listener := range *listeners {
-		if listener.closed {
-			continue
-		}
-
-		if err := listener.callback(event); err != nil {
-			listener.errors <- err
-			listener.closed = true
-		}
-	}
-	*log = append(*log, event)
-
-	lock.Unlock()
-}
-
-func (ev *eventHandler) logEvent(event *proto.Event) {
-	ev.log(event, &ev.eventListeners, &ev.eventLog, &ev.logLock)
-}
-
-func (ev *eventHandler) logApplicationLog(event *proto.Event) {
-	ev.log(event, &ev.applicationLogListeners, &ev.applicationLogs, &ev.applicationLogsLock)
+func (ev *eventHandler) forEachApplicationLog(callback func(*proto.Event) error) error {
+	return ev.forEach(&ev.applicationLogListeners, &ev.applicationLogs, &ev.applicationLogsLock, callback)
 }
 
 func (ev *eventHandler) forEach(listeners *[]*listener, log *[]*proto.Event, lock sync.Locker, callback func(*proto.Event) error) error {
@@ -177,169 +136,42 @@ func (ev *eventHandler) forEach(listeners *[]*listener, log *[]*proto.Event, loc
 	return <-listener.errors
 }
 
-func (ev *eventHandler) forEachEvent(callback func(*proto.Event) error) error {
-	select {
-	case handler.wait <- true:
-	default:
+func Handle(event *proto.Event) error {
+	if event != nil {
+		handler.handle(event)
 	}
-	return ev.forEach(&ev.eventListeners, &ev.eventLog, &ev.logLock, callback)
+	return nil
 }
 
-func (ev *eventHandler) forEachApplicationLog(callback func(*proto.Event) error) error {
-	return ev.forEach(&ev.applicationLogListeners, &ev.applicationLogs, &ev.applicationLogsLock, callback)
+// WaitForConnection will block execution until the server receives a connection
+func WaitForConnection() {
+	<-handler.wait
 }
 
-func emptyState(cfg Config) *proto.State {
-	builds := map[string]string{}
-	for _, p := range cfg.GetPipelines() {
-		for _, a := range p.Build.Artifacts {
-			builds[a.ImageName] = NotStarted
+func (ev *eventHandler) logEvent(event *proto.Event) {
+	ev.log(event, &ev.eventListeners, &ev.eventLog, &ev.logLock)
+}
+
+func (ev *eventHandler) logApplicationLog(event *proto.Event) {
+	ev.log(event, &ev.applicationLogListeners, &ev.applicationLogs, &ev.applicationLogsLock)
+}
+
+func (ev *eventHandler) log(event *proto.Event, listeners *[]*listener, log *[]*proto.Event, lock sync.Locker) {
+	lock.Lock()
+
+	for _, listener := range *listeners {
+		if listener.closed {
+			continue
+		}
+
+		if err := listener.callback(event); err != nil {
+			listener.errors <- err
+			listener.closed = true
 		}
 	}
-	metadata := initializeMetadata(cfg.GetPipelines(), cfg.GetKubeContext(), cfg.GetRunID())
-	return emptyStateWithArtifacts(builds, metadata, cfg.AutoBuild(), cfg.AutoDeploy(), cfg.AutoSync())
-}
+	*log = append(*log, event)
 
-func emptyStateWithArtifacts(builds map[string]string, metadata *proto.Metadata, autoBuild, autoDeploy, autoSync bool) *proto.State {
-	return &proto.State{
-		BuildState: &proto.BuildState{
-			Artifacts:   builds,
-			AutoTrigger: autoBuild,
-			StatusCode:  proto.StatusCode_OK,
-		},
-		TestState: &proto.TestState{
-			Status:     NotStarted,
-			StatusCode: proto.StatusCode_OK,
-		},
-		RenderState: &proto.RenderState{
-			Status:     NotStarted,
-			StatusCode: proto.StatusCode_OK,
-		},
-		DeployState: &proto.DeployState{
-			Status:      NotStarted,
-			AutoTrigger: autoDeploy,
-			StatusCode:  proto.StatusCode_OK,
-		},
-		StatusCheckState: emptyStatusCheckState(),
-		ForwardedPorts:   make(map[int32]*proto.PortForwardEvent),
-		FileSyncState: &proto.FileSyncState{
-			Status:      NotStarted,
-			AutoTrigger: autoSync,
-		},
-		Metadata: metadata,
-	}
-}
-
-// ResetStateOnBuild resets the build, test, deploy and sync state
-func ResetStateOnBuild() {
-	builds := map[string]string{}
-	for k := range handler.getState().BuildState.Artifacts {
-		builds[k] = NotStarted
-	}
-	autoBuild, autoDeploy, autoSync := handler.getState().BuildState.AutoTrigger, handler.getState().DeployState.AutoTrigger, handler.getState().FileSyncState.AutoTrigger
-	newState := emptyStateWithArtifacts(builds, handler.getState().Metadata, autoBuild, autoDeploy, autoSync)
-	handler.setState(newState)
-}
-
-// ResetStateOnTest resets the test, deploy, sync and status check state
-func ResetStateOnTest() {
-	newState := handler.getState()
-	newState.TestState.Status = NotStarted
-	handler.setState(newState)
-}
-
-// ResetStateOnDeploy resets the deploy, sync and status check state
-func ResetStateOnDeploy() {
-	newState := handler.getState()
-	newState.DeployState.Status = NotStarted
-	newState.DeployState.StatusCode = proto.StatusCode_OK
-	newState.StatusCheckState = emptyStatusCheckState()
-	newState.ForwardedPorts = map[int32]*proto.PortForwardEvent{}
-	newState.DebuggingContainers = nil
-	handler.setState(newState)
-}
-
-func UpdateStateAutoBuildTrigger(t bool) {
-	newState := handler.getState()
-	newState.BuildState.AutoTrigger = t
-	handler.setState(newState)
-}
-
-func UpdateStateAutoDeployTrigger(t bool) {
-	newState := handler.getState()
-	newState.DeployState.AutoTrigger = t
-	handler.setState(newState)
-}
-
-func UpdateStateAutoSyncTrigger(t bool) {
-	newState := handler.getState()
-	newState.FileSyncState.AutoTrigger = t
-	handler.setState(newState)
-}
-
-func emptyStatusCheckState() *proto.StatusCheckState {
-	return &proto.StatusCheckState{
-		Status:     NotStarted,
-		Resources:  map[string]string{},
-		StatusCode: proto.StatusCode_OK,
-	}
-}
-
-// InitializeState instantiates the global state of the skaffold runner, as well as the event log.
-func InitializeState(cfg Config) {
-	handler.cfg = cfg
-	handler.setState(emptyState(cfg))
-}
-
-func AutoTriggerDiff(phase constants.Phase, val bool) (bool, error) {
-	switch phase {
-	case constants.Build:
-		return val != handler.getState().BuildState.AutoTrigger, nil
-	case constants.Sync:
-		return val != handler.getState().FileSyncState.AutoTrigger, nil
-	case constants.Deploy:
-		return val != handler.getState().DeployState.AutoTrigger, nil
-	default:
-		return false, fmt.Errorf("unknown Phase %v not found in handler state", phase)
-	}
-}
-
-func TaskInProgress(task constants.Phase, description string) {
-	// Special casing to increment iteration and clear application and skaffold logs
-	if task == constants.DevLoop {
-		handler.iteration++
-
-		handler.applicationLogs = []*proto.Event{}
-	}
-
-	handler.handleTaskEvent(&proto.TaskEvent{
-		Id:          fmt.Sprintf("%s-%d", task, handler.iteration),
-		Task:        string(task),
-		Description: description,
-		Iteration:   int32(handler.iteration),
-		Status:      InProgress,
-	})
-}
-
-func TaskFailed(task constants.Phase, err error) {
-	ae := sErrors.ActionableErrV2(handler.cfg, task, err)
-	handler.sendErrorMessage(task, constants.SubtaskIDNone, err)
-	handler.handleTaskEvent(&proto.TaskEvent{
-		Id:            fmt.Sprintf("%s-%d", task, handler.iteration),
-		Task:          string(task),
-		Iteration:     int32(handler.iteration),
-		Status:        Failed,
-		ActionableErr: ae,
-	})
-}
-
-func TaskSucceeded(task constants.Phase) {
-	handler.handleTaskEvent(&proto.TaskEvent{
-		Id:        fmt.Sprintf("%s-%d", task, handler.iteration),
-		Task:      string(task),
-		Iteration: int32(handler.iteration),
-		Status:    Succeeded,
-	})
+	lock.Unlock()
 }
 
 // PortForwarded notifies that a remote port has been forwarded locally.
@@ -384,28 +216,13 @@ func (ev *eventHandler) sendErrorMessage(task constants.Phase, subtask string, e
 	})
 }
 
-func (ev *eventHandler) setState(state *proto.State) {
-	ev.stateLock.Lock()
-	ev.state = state
-	ev.stateLock.Unlock()
-}
-
 func (ev *eventHandler) handle(event *proto.Event) {
 	event.Timestamp = timestamppb.Now()
 	ev.eventChan <- event
 	if _, ok := event.GetEventType().(*proto.Event_TerminationEvent); ok {
-		// close the event channel indicating there are no more events to all the
-		// receivers
+		// close the event channel indicating there are no more events to all the receivers
 		close(ev.eventChan)
 	}
-}
-
-func (ev *eventHandler) handleTaskEvent(e *proto.TaskEvent) {
-	ev.handle(&proto.Event{
-		EventType: &proto.Event_TaskEvent{
-			TaskEvent: e,
-		},
-	})
 }
 
 func (ev *eventHandler) handleExec(event *proto.Event) {
