+++ pkg/skaffold/event/v2/event_test.go
@@ -27,10 +27,7 @@ import (
 
 	//nolint:golint,staticcheck
 	"github.com/golang/protobuf/jsonpb"
-	"github.com/google/go-cmp/cmp/cmpopts"
-	"google.golang.org/protobuf/testing/protocmp"
 
-	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/constants"
 	latestV1 "github.com/GoogleContainerTools/skaffold/pkg/skaffold/schema/latest/v1"
 	proto "github.com/GoogleContainerTools/skaffold/proto/v2"
 	"github.com/GoogleContainerTools/skaffold/testutil"
@@ -76,19 +73,6 @@ func TestGetLogEvents(t *testing.T) {
 	}
 }
 
-func TestGetState(t *testing.T) {
-	ev := newHandler()
-	ev.state = emptyState(mockCfg([]latestV1.Pipeline{{}}, "test"))
-
-	ev.stateLock.Lock()
-	ev.state.BuildState.Artifacts["img"] = Complete
-	ev.stateLock.Unlock()
-
-	state := ev.getState()
-
-	testutil.CheckDeepEqual(t, Complete, state.BuildState.Artifacts["img"])
-}
-
 func wait(t *testing.T, condition func() bool) {
 	ticker := time.NewTicker(10 * time.Millisecond)
 	defer ticker.Stop()
@@ -106,248 +90,6 @@ func wait(t *testing.T, condition func() bool) {
 		case <-timeout.C:
 			t.Fatal("Timed out waiting")
 		}
-	}
-}
-
-func TestResetStateOnBuild(t *testing.T) {
-	defer func() { handler = newHandler() }()
-	handler = newHandler()
-	handler.state = &proto.State{
-		BuildState: &proto.BuildState{
-			Artifacts: map[string]string{
-				"image1": Complete,
-			},
-		},
-		RenderState: &proto.RenderState{Status: Complete},
-		DeployState: &proto.DeployState{Status: Complete},
-		ForwardedPorts: map[int32]*proto.PortForwardEvent{
-			2001: {
-				LocalPort:  2000,
-				PodName:    "test/pod",
-				TargetPort: &targetPort,
-			},
-		},
-		StatusCheckState: &proto.StatusCheckState{Status: Complete},
-		FileSyncState:    &proto.FileSyncState{Status: Succeeded},
-	}
-
-	ResetStateOnBuild()
-	expected := &proto.State{
-		BuildState: &proto.BuildState{
-			Artifacts: map[string]string{
-				"image1": NotStarted,
-			},
-		},
-		TestState:        &proto.TestState{Status: NotStarted},
-		RenderState:      &proto.RenderState{Status: NotStarted},
-		DeployState:      &proto.DeployState{Status: NotStarted},
-		StatusCheckState: &proto.StatusCheckState{Status: NotStarted, Resources: map[string]string{}},
-		FileSyncState:    &proto.FileSyncState{Status: NotStarted},
-	}
-	testutil.CheckDeepEqual(t, expected, handler.getState(), cmpopts.EquateEmpty(), protocmp.Transform())
-}
-
-func TestResetStateOnDeploy(t *testing.T) {
-	defer func() { handler = newHandler() }()
-	handler = newHandler()
-	handler.state = &proto.State{
-		BuildState: &proto.BuildState{
-			Artifacts: map[string]string{
-				"image1": Complete,
-			},
-		},
-		DeployState: &proto.DeployState{Status: Complete},
-		ForwardedPorts: map[int32]*proto.PortForwardEvent{
-			2001: {
-				LocalPort:  2000,
-				PodName:    "test/pod",
-				TargetPort: &targetPort,
-			},
-		},
-		StatusCheckState: &proto.StatusCheckState{Status: Complete},
-	}
-	ResetStateOnDeploy()
-	expected := &proto.State{
-		BuildState: &proto.BuildState{
-			Artifacts: map[string]string{
-				"image1": Complete,
-			},
-		},
-		DeployState: &proto.DeployState{Status: NotStarted},
-		StatusCheckState: &proto.StatusCheckState{Status: NotStarted,
-			Resources: map[string]string{},
-		},
-	}
-	testutil.CheckDeepEqual(t, expected, handler.getState(), cmpopts.EquateEmpty(), protocmp.Transform())
-}
-
-func TestEmptyStateCheckState(t *testing.T) {
-	actual := emptyStatusCheckState()
-	expected := &proto.StatusCheckState{Status: NotStarted,
-		Resources: map[string]string{},
-	}
-	testutil.CheckDeepEqual(t, expected, actual, cmpopts.EquateEmpty(), protocmp.Transform())
-}
-
-func TestUpdateStateAutoTriggers(t *testing.T) {
-	defer func() { handler = newHandler() }()
-	handler = newHandler()
-	handler.state = &proto.State{
-		BuildState: &proto.BuildState{
-			Artifacts: map[string]string{
-				"image1": Complete,
-			},
-			AutoTrigger: false,
-		},
-		DeployState: &proto.DeployState{Status: Complete, AutoTrigger: false},
-		ForwardedPorts: map[int32]*proto.PortForwardEvent{
-			2001: {
-				LocalPort:  2000,
-				PodName:    "test/pod",
-				TargetPort: &targetPort,
-			},
-		},
-		StatusCheckState: &proto.StatusCheckState{Status: Complete},
-		FileSyncState: &proto.FileSyncState{
-			Status:      "Complete",
-			AutoTrigger: false,
-		},
-	}
-	UpdateStateAutoBuildTrigger(true)
-	UpdateStateAutoDeployTrigger(true)
-	UpdateStateAutoSyncTrigger(true)
-
-	expected := &proto.State{
-		BuildState: &proto.BuildState{
-			Artifacts: map[string]string{
-				"image1": Complete,
-			},
-			AutoTrigger: true,
-		},
-		DeployState: &proto.DeployState{Status: Complete, AutoTrigger: true},
-		ForwardedPorts: map[int32]*proto.PortForwardEvent{
-			2001: {
-				LocalPort:  2000,
-				PodName:    "test/pod",
-				TargetPort: &targetPort,
-			},
-		},
-		StatusCheckState: &proto.StatusCheckState{Status: Complete},
-		FileSyncState: &proto.FileSyncState{
-			Status:      "Complete",
-			AutoTrigger: true,
-		},
-	}
-	testutil.CheckDeepEqual(t, expected, handler.getState(), cmpopts.EquateEmpty(), protocmp.Transform())
-}
-
-func TestTaskFailed(t *testing.T) {
-	tcs := []struct {
-		description string
-		state       *proto.State
-		phase       constants.Phase
-		waitFn      func() bool
-	}{
-		{
-			description: "build failed",
-			phase:       constants.Build,
-			waitFn: func() bool {
-				handler.logLock.Lock()
-				logEntry := handler.eventLog[len(handler.eventLog)-1]
-				handler.logLock.Unlock()
-				te := logEntry.GetTaskEvent()
-				return te != nil && te.Status == Failed && te.Id == "Build-0"
-			},
-		},
-		{
-			description: "deploy failed",
-			phase:       constants.Deploy,
-			waitFn: func() bool {
-				handler.logLock.Lock()
-				logEntry := handler.eventLog[len(handler.eventLog)-1]
-				handler.logLock.Unlock()
-				te := logEntry.GetTaskEvent()
-				return te != nil && te.Status == Failed && te.Id == "Deploy-0"
-			},
-		},
-		{
-			description: "status check failed",
-			phase:       constants.StatusCheck,
-			waitFn: func() bool {
-				handler.logLock.Lock()
-				logEntry := handler.eventLog[len(handler.eventLog)-1]
-				handler.logLock.Unlock()
-				te := logEntry.GetTaskEvent()
-				return te != nil && te.Status == Failed && te.Id == "StatusCheck-0"
-			},
-		},
-	}
-	for _, tc := range tcs {
-		t.Run(tc.description, func(t *testing.T) {
-			TaskFailed(tc.phase, errors.New("random error"))
-			wait(t, tc.waitFn)
-		})
-	}
-}
-
-func TestAutoTriggerDiff(t *testing.T) {
-	tests := []struct {
-		description  string
-		phase        constants.Phase
-		handlerState *proto.State
-		val          bool
-		expected     bool
-	}{
-		{
-			description: "build needs update",
-			phase:       constants.Build,
-			val:         true,
-			handlerState: &proto.State{
-				BuildState: &proto.BuildState{
-					AutoTrigger: false,
-				},
-			},
-			expected: true,
-		},
-		{
-			description: "deploy doesn't need update",
-			phase:       constants.Deploy,
-			val:         true,
-			handlerState: &proto.State{
-				BuildState: &proto.BuildState{
-					AutoTrigger: false,
-				},
-				DeployState: &proto.DeployState{
-					AutoTrigger: true,
-				},
-			},
-			expected: false,
-		},
-		{
-			description: "sync needs update",
-			phase:       constants.Sync,
-			val:         false,
-			handlerState: &proto.State{
-				FileSyncState: &proto.FileSyncState{
-					AutoTrigger: true,
-				},
-			},
-			expected: true,
-		},
-	}
-
-	for _, test := range tests {
-		testutil.Run(t, test.description, func(t *testutil.T) {
-			// Setup handler state
-			handler.setState(test.handlerState)
-
-			got, err := AutoTriggerDiff(test.phase, test.val)
-			if err != nil {
-				t.Fail()
-			}
-
-			t.CheckDeepEqual(test.expected, got)
-		})
 	}
 }
